# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medcom API",
    "description": "Medcom API documentation",
    "termsOfService": "",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/pharmacies/{id}": {
      "get": {
        "tags": ["pharmacy-resource"],
        "operationId": "getPharmacy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["pharmacy-resource"],
        "operationId": "updatePharmacy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDTO"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["pharmacy-resource"],
        "operationId": "partialUpdatePharmacy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDTO"
              }
            },
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{login}": {
      "get": {
        "tags": ["user-resource"],
        "operationId": "getUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["user-resource"],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user-resource"],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": ["user-resource"],
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["user-resource"],
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["public-user-resource"],
        "operationId": "getAllPublicUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PharmacyDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string"
          }
        }
      },
...
      "AdminUserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
          },
          "firstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 256,
            "minLength": 0
          },
          "activated": {
            "type": "boolean"
          },
          "langKey": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2
          },
          "createdBy": {
            "type": "string",
            "readOnly": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "lastModifiedBy": {
            "type": "string",
            "readOnly": true
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "pharmacyId": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "required": ["login"]
      },
...
      "UserCreationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
          },
          "firstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 256,
            "minLength": 0
          },
          "activated": {
            "type": "boolean"
          },
          "langKey": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": ["login"]
      },
      "Pharmacy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "createdBy": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "string"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
          },
          "firstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "activated": {
            "type": "boolean"
          },
          "langKey": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 256,
            "minLength": 0
          },
          "resetDate": {
            "type": "string",
            "format": "date-time"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/Pharmacy"
          }
        },
        "required": ["activated", "login"]
      },
...
    }
  }
}